Refactor: #73 - search 페이지 컴포넌트 분리
